There appears to be a bug in Spring/Java where @CookieValue terminates abruptly if given a cookie with a '(' in it e.g. from Javascript. If you use javascript to set a cookie       '20% semicolon; code is something !@#$%^&*() test encode' but the @CookieValue only sees       '20% semicolon; code is something !@#$%^&*'' losing everything after the braces.
The Chrome request debugger shows the browser has sent it correctly to the server, with the '(' and ')' unencoded, it seems Java is truncating it before it gets to the Spring layer (using the debugger on the servlet spec's HttpRequest shows its been truncated there).

Workaround is to explicitly encode the '(' and ')' characters, since encodeUriComponent in javascript doesn't encode them.